requirements.txt
pyodbc
requests
python-dotenv
tk
